+++ src/Network/JotForm/Core.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 module Network.JotForm.Core
     ( -- * API Client
       ApiClient (..)
@@ -36,10 +38,8 @@ module Network.JotForm.Core
       -- ** JsonException
     , JsonException (..)
 
-      -- * Constants
-    , userAgent
-
       -- * Headers
+    , userAgent
     , urlEncode
     , acceptJson
     ) where
@@ -47,18 +47,18 @@ module Network.JotForm.Core
 import Control.Exception (Exception, throwIO)
 import Data.Aeson (FromJSON)
 import Data.Aeson qualified as Json
-import Data.ByteString qualified as Byte.Str
-import Data.ByteString qualified as Str (ByteString)
-import Data.ByteString.Char8 qualified as Byte.Str.Char8
 import Data.ByteString.Lazy qualified as Lazy (ByteString)
+import Data.Text qualified as Str (Text)
+import Data.Text qualified as Text.Str
+import Data.Text.Encoding qualified as Text.Str.Enc
 import Network.HTTP.Client (Manager, Request, Response)
 import Network.HTTP.Client qualified as Client
 import Network.HTTP.Client.TLS qualified as Client.TLS
-import Network.HTTP.Types (Header, Method, Query)
+import Network.HTTP.Types (Header, Method, QueryText)
 import Network.HTTP.Types.Header qualified as Header
 import Network.JotForm.Utils qualified as Utils
 
-newtype JsonException = MkJsonException String
+newtype JsonException = MkJsonException Str.Text
     deriving (Eq, Ord, Show, Read)
 
 instance Exception JsonException
@@ -74,35 +74,31 @@ data DebugMode = DebugOn | DebugOff
 
 data ApiClient = MkApiClient
     { baseUrl :: BaseUrl
-    , apiVersion :: Str.ByteString
-    , apiKey :: Str.ByteString
+    , apiVersion :: Str.Text
+    , apiKey :: Str.Text
     , outputType :: OutputType
     , debugMode :: DebugMode
     , httpManager :: Manager
     }
 
 data Params = MkParams
-    { path :: Str.ByteString
-    , query :: Query
-    , body :: Str.ByteString
+    { path :: Str.Text
+    , query :: QueryText
+    , body :: Str.Text
     , headers :: [Header]
     , method :: Method
     }
     deriving (Eq, Ord, Read, Show)
 
-defaultParams :: Str.ByteString -> Method -> Params
+defaultParams :: Str.Text -> Method -> Params
 defaultParams thisPath thisMethod =
     MkParams
         { path = thisPath
         , query = []
-        , body = Byte.Str.empty
+        , body = Text.Str.empty
         , headers = []
         , method = thisMethod
         }
-
--- | The User-Agent which is used for all the requests made to JotForm.
-userAgent :: Str.ByteString
-userAgent = Utils.ascii "JOTFORM_HASKELL_WRAPPER"
 
 defaultSecureRequest :: Request
 defaultSecureRequest =
@@ -115,18 +111,18 @@ defaultSecureRequest =
 -- TLS-enabled manager config from @http-client-tls@.
 --
 -- To pass a different 'Manager' instead use the 'defaultApiClient'' function.
-defaultApiClient :: Str.ByteString -> IO ApiClient
+defaultApiClient :: Str.Text -> IO ApiClient
 defaultApiClient key = do
     manager <- Client.newManager Client.TLS.tlsManagerSettings
     pure $ defaultApiClient' key manager
 
 -- | Creates an 'ApiClient' using default settings, and the given
 -- HTTP 'Manager'.
-defaultApiClient' :: Str.ByteString -> Manager -> ApiClient
+defaultApiClient' :: Str.Text -> Manager -> ApiClient
 defaultApiClient' key manager =
     MkApiClient
         { baseUrl = DefaultBaseUrl
-        , apiVersion = Utils.ascii "v1"
+        , apiVersion = "v1"
         , apiKey = key
         , outputType = JsonOutput
         , debugMode = DebugOff
@@ -135,22 +131,22 @@ defaultApiClient' key manager =
 
 -- | The same as 'defaultApiClient', but is set to use the EU endpoint - use
 -- this if your account is in EU Safe mode.
-defaultApiClientEu :: Str.ByteString -> IO ApiClient
+defaultApiClientEu :: Str.Text -> IO ApiClient
 defaultApiClientEu key = do
     def <- defaultApiClient key
     pure $ def {baseUrl = EuBaseUrl}
 
 -- | The same as 'defaultApiClient'', but is set to use the EU endpoint - use
 -- this if your account is in EU Safe mode.
-defaultApiClientEu' :: Str.ByteString -> Manager -> ApiClient
+defaultApiClientEu' :: Str.Text -> Manager -> ApiClient
 defaultApiClientEu' key manager =
     let def = defaultApiClient' key manager
     in  def {baseUrl = EuBaseUrl}
 
-baseUrlToString :: BaseUrl -> Str.ByteString
+baseUrlToString :: BaseUrl -> Str.Text
 baseUrlToString = \case
-    DefaultBaseUrl -> Utils.ascii "api.jotform.com"
-    EuBaseUrl -> Utils.ascii "eu-api.jotform.com"
+    DefaultBaseUrl -> "api.jotform.com"
+    EuBaseUrl -> "eu-api.jotform.com"
 
 fetch :: ApiClient -> Params -> IO (Response Lazy.ByteString)
 fetch client params = Client.httpLbs request manager
@@ -162,7 +158,7 @@ fetchJson :: FromJSON a => ApiClient -> Params -> IO (Response a)
 fetchJson client params = do
     response <- Client.httpLbs requestJson (httpManager client)
     case Json.eitherDecode $ Client.responseBody response of
-        Left err -> throwIO $ MkJsonException err
+        Left err -> throwIO $ MkJsonException $ Text.Str.pack err
         Right json -> pure (json <$ response)
   where
     request = toRequest client params
@@ -171,24 +167,28 @@ fetchJson client params = do
 toRequest :: ApiClient -> Params -> Request
 toRequest client params =
     defaultSecureRequest
-        { Client.host = baseUrlToString $ baseUrl client
-        , Client.path = versionPath <> path params <> outputPath
+        { Client.host = encode $ baseUrlToString $ baseUrl client
+        , Client.path = encode $ versionPath <> path params <> outputPath
         , Client.method = method params
-        , Client.queryString = Utils.renderQuery $ query params
-        , Client.requestBody = Client.RequestBodyBS $ body params
+        , Client.queryString = Utils.renderQueryBytes $ query params
+        , Client.requestBody = Client.RequestBodyBS $ encode $ body params
         , Client.requestHeaders = headers params <> defHeaders
         }
   where
-    versionPath = Byte.Str.Char8.cons '/' $ apiVersion client
-    outputPath = case outputType client of JsonOutput -> Byte.Str.empty
+    encode = Text.Str.Enc.encodeUtf8
+    versionPath = Text.Str.cons '/' $ apiVersion client
+    outputPath = case outputType client of JsonOutput -> Text.Str.empty
     defHeaders =
-        [ (Utils.headerName "ApiKey", apiKey client)
-        , (Header.hUserAgent, userAgent)
+        [ (Utils.headerName "ApiKey", encode $ apiKey client)
+        , userAgent
         ]
 
+-- | The User-Agent which is used for all the requests made to JotForm.
+userAgent :: Header
+userAgent = (Header.hUserAgent, "JOTFORM_HASKELL_WRAPPER")
+
 urlEncode :: Header
-urlEncode =
-    (Header.hContentType, Utils.ascii "application/x-www-form-urlencoded")
+urlEncode = (Header.hContentType, "application/x-www-form-urlencoded")
 
 acceptJson :: Header
-acceptJson = (Header.hAccept, Utils.ascii "application/json")
+acceptJson = (Header.hAccept, "application/json")