+++ src/Network/JotForm/Utils.hs
@@ -1,31 +1,51 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 module Network.JotForm.Utils
-    ( headerName
-    , ascii
-    , showAscii
+    ( -- * Headers
+      headerName
     , updateHeaders
-    , encodeStrict
-    , renderQuery
+
+      -- * Text
+    , showText
+    , elemIndexText
+
+      -- * JSON
+    , encodeText
+
+      -- * Date / Time
+    , formatDateText
+    , showDateJF
+
+      -- * QueryText
+    , renderQueryText
+    , renderQueryBytes
     ) where
 
 import Data.Aeson (ToJSON)
-import Data.Aeson qualified as Json
+import Data.Aeson.Text qualified as Json.Text
 import Data.ByteString qualified as Byte.Str
 import Data.ByteString qualified as Str (ByteString)
-import Data.ByteString.Char8 qualified as Byte.Str.Char8
+import Data.ByteString.Builder qualified as Byte.Lazy.Bldr
 import Data.CaseInsensitive qualified as CaseIns
+import Data.Text qualified as Str (Text)
+import Data.Text qualified as Text.Str
+import Data.Text.Encoding qualified as Text.Str.Enc
+import Data.Text.Lazy qualified as Text.Lazy
+import Data.Time (FormatTime, TimeLocale, Day)
+import Data.Time qualified as Time
 import Network.HTTP.Client (Request)
 import Network.HTTP.Client qualified as Client
-import Network.HTTP.Types (HeaderName, Query, RequestHeaders)
+import Network.HTTP.Types (HeaderName, QueryText, RequestHeaders)
 import Network.HTTP.Types.URI qualified as URI
 
-headerName :: String -> HeaderName
-headerName = CaseIns.mk . ascii
+headerName :: Str.Text -> HeaderName
+headerName = CaseIns.mk . Text.Str.Enc.encodeUtf8
 
-ascii :: String -> Str.ByteString
-ascii = Byte.Str.Char8.pack
+elemIndexText :: Char -> Str.Text -> Maybe Int
+elemIndexText c = Text.Str.findIndex (== c)
 
-showAscii :: Show a => a -> Str.ByteString
-showAscii = ascii . show
+showText :: Show a => a -> Str.Text
+showText = Text.Str.pack . show
 
 updateHeaders :: (RequestHeaders -> RequestHeaders) -> Request -> Request
 updateHeaders modify request =
@@ -34,10 +54,23 @@ updateHeaders modify request =
             modify $ Client.requestHeaders request
         }
 
-encodeStrict :: ToJSON a => a -> Str.ByteString
-encodeStrict = Byte.Str.toStrict . Json.encode
+encodeText :: ToJSON a => a -> Str.Text
+encodeText = Text.Lazy.toStrict . Json.Text.encodeToLazyText
 
 -- | Convenience function - as we never need to render the '?' at the start
 -- of the query, it is easier to not have to specify that at every call site.
-renderQuery :: Query -> Str.ByteString
-renderQuery = URI.renderQuery False
+renderQueryText :: QueryText -> Str.Text
+renderQueryText = Text.Str.Enc.decodeUtf8 . renderQueryBytes
+
+renderQueryBytes :: QueryText -> Str.ByteString
+renderQueryBytes =
+    Byte.Str.toStrict
+        . Byte.Lazy.Bldr.toLazyByteString
+        . URI.renderQueryText False
+
+formatDateText :: FormatTime t => TimeLocale -> Str.Text -> t -> Str.Text
+formatDateText locale fmt time =
+    Text.Str.pack $ Time.formatTime locale (Text.Str.unpack fmt) time
+
+showDateJF :: Day -> Str.Text
+showDateJF = formatDateText Time.defaultTimeLocale "%m/%d/%Y"