+++ test/Examples.hs
@@ -7,10 +7,11 @@ import Data.Aeson qualified as Json
 import Data.Aeson.Encode.Pretty qualified as Json.Pretty
 import Data.Aeson.Optics (key, _Integer, _String)
 import Data.ByteString qualified as Byte.Str
-import Data.ByteString qualified as Str (ByteString)
-import Data.ByteString.Char8 qualified as Byte.Str.Char8
 import Data.ByteString.Lazy.Char8 qualified as Byte.Lazy.Char8
 import Data.Foldable (for_)
+import Data.Text qualified as Str (Text)
+import Data.Text qualified as Text.Str
+import Data.Text.Encoding qualified as Text.Str.Enc
 import Network.JotForm (ApiClient)
 import Network.JotForm qualified as JotForm
 import Optics.Core ((%), (^?))
@@ -80,10 +81,10 @@ getClient =
     getClientFrom "test-api-key-eu.txt" JotForm.defaultApiClientEu
         <|> getClientFrom "test-api-key.txt" JotForm.defaultApiClient
 
-getClientFrom :: FilePath -> (Str.ByteString -> IO ApiClient) -> IO ApiClient
+getClientFrom :: FilePath -> (Str.Text -> IO ApiClient) -> IO ApiClient
 getClientFrom fileName mkClient = do
-    apiKey <- Byte.Str.readFile fileName
-    mkClient $ Byte.Str.Char8.strip apiKey
+    apiKey <- Text.Str.Enc.decodeUtf8 <$> Byte.Str.readFile fileName
+    mkClient $ Text.Str.strip apiKey
 
 main :: IO ()
 main = do