+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for jotform-api-haskell
 
+## 0.1.1.0 -- 2023-06-04
+
+* use proper types for main interfaces (including Text and Day), converting to
+  ByteString when we make the requests
+
 ## 0.1.0.0 -- 2023-05-17
 
 * First release.